+++ NEWS.md
@@ -1,10 +1,12 @@
-# ghypernet 1.0.1.0921
+# ghypernet 1.1.0
 
 - function renaming:
 * `linkSignificance()` --> `link_significance()`
 * `ComputeXi()` --> `compute_xi()`
 
 - new features
+* updated documentation for endogenous statistics functions
+* introduced a new function `get_zero_dummy()` that generates the dummy variable to be used together with endogenous statistics functions.
 * added testing with test_that
 * `check_specs()`: recognize bipartite graphs
 * addressed issue #15: bccm now names coefficients according to label names. E.g.: coefficient for the propensities between block 'red' and 'blue' are now named 'red<->blue'.